… tensorflow/compiler/jit/compilability_check_util.cc
… tensorflow/compiler/jit/compilability_check_util.h
… tensorflow/compiler/jit/mark_for_compilation_pass.cc
… tensorflow/compiler/tf2xla/BUILD
… tensorflow/compiler/tf2xla/functionalize_cond.cc
… tensorflow/compiler/tf2xla/functionalize_while.cc
… tensorflow/compiler/xla/service/gpu/kernel_thunk.cc
… tensorflow/compiler/xla/service/gpu/launch_dimensions.cc
… tensorflow/compiler/xla/service/gpu/launch_dimensions.h
… tensorflow/compiler/xla/service/gpu/parallel_loop_emitter.cc
… tensorflow/compiler/xla/service/gpu/stream_executor_util.cc
… tensorflow/compiler/xla/service/gpu/stream_executor_util.h
… tensorflow/compiler/xla/service/gpu/tests/BUILD
… tensorflow/compiler/xla/service/gpu/tests/parallel_reduction_test.cc
… tensorflow/compiler/xla/union_find.h