Extend the XLA codegen to generate parallel reductions when there are multiple
reduce instructions in a fusion computation. This paves a way for horizontal input
fusion in the future.

### Performance Numbers
Now we see ~3% e2e gain for NVIDIA JoC BERT.

For `ManyParallelReductions` with 128 reduce instructions in the unittest, the
execution time is reduced from 325us to 3.9us (83X), reported by nvprof as below.

Before:

            Type  Time(%)      Time     Calls       Avg       Min       Max  Name

                   32.50%  325.54us         1  325.54us  325.54us  325.54us  fusion

After:

            Type  Time(%)      Time     Calls       Avg       Min       Max  Name

                    0.59%  3.9030us         1  3.9030us  3.9030us  3.9030us  fusion


### High-level descriptions about the Algorithm
The overall idea is to divide the reduce instructions in a fusion into groups for parallelization, as implemented by `DivideOutputInstructionsIntoGroups()`. Different groups will be executed in parallel, while reduce instructions in the same group are run sequentially. Generally speaking, we'd like to run the reduce instructions in parallel without incurring too much re-computation overhead. The current heuristic is to place reduce instructions who share nothing or only
 (broadcasted) scalars/constants into different groups; otherwise, they are placed in the same group. Non-reduce instructions always go with the reduce instructions into the same group so long as they share any predecessors.


To avoid complicating the (already complicated) index calculation in reduce code generation, we use raw (CUDA) block_id_y to select the reduce groups (instead of thread_id_x/block_id_x). That is, a block_id_y is assigned to a group and so different groups can be run in parallel. Separation of the reduce codegen index calculation (using raw dim x) and parallelization (using raw dim y) should also provide better future maintainability.