@cheshire 2020-08-28
Thanks! I'm fine with this as long as nitpicks are addressed.