@trentlo 2020-08-31
> @trentlo I'm fine with merging this.

Great! I do not see regressions from the NV-internal benchmarks but we will see.