@trentlo 2020-08-31
George, could you launch the CI again? Thanks!

There were build errors due to narrowing type conversion (In MacOS CPU Python3 and Bazel Windows GPU). I fixed them by pushing a new commit. An error in Bazel Windows was not seem to related to the PR. Other tests were clean.