@cheshire 2020-09-01
(we have a long-standing bug to enable these warnings in OSS, but so far it hasn't been done =/)