@trentlo 2020-09-01
> @trentlo Could you fix the following warnings?
> 
> ```
> third_party/tensorflow/compiler/xla/service/gpu/ir_emitter_unnested.cc:3007:26: error: non-constant-expression cannot be narrowed from type 'size_t' (aka 'unsigned long') to 'long long' in initializer list [-Wc++11-narrowing]
>       return ShapeIndex({i});
>                          ^
> third_party/tensorflow/compiler/xla/service/gpu/ir_emitter_unnested.cc:3007:26: note: insert an explicit cast to silence this issue
>       return ShapeIndex({i});
>                          ^
>                          static_cast<long long>( )
> third_party/tensorflow/compiler/xla/service/gpu/ir_emitter_unnested.cc:3834:50: error: '&&' within '||' [-Werror,-Wlogical-op-parentheses]
>          HloOpcode::kBroadcast == instr.opcode() &&
>          ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~^~
> third_party/tensorflow/compiler/xla/service/gpu/ir_emitter_unnested.cc:3834:50: note: place parentheses around the '&&' expression to silence this warning
>          HloOpcode::kBroadcast == instr.opcode() &&
>                                                  ^
>          (
> third_party/tensorflow/compiler/xla/service/gpu/ir_emitter_unnested.cc:3957:7: error: ignoring return value of function declared with 'warn_unused_result' attribute [-Werror,-Wunused-result]
>       EmitIRForReduction(unnested_hlo, instr_groups[i], &reduction_info,
>       ^~~~~~~~~~~~~~~~~~ ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
> third_party/tensorflow/compiler/xla/service/gpu/ir_emitter_unnested.cc:3979:56: error: non-constant-expression cannot be narrowed from type 'std::__u::vector<std::__u::vector<xla::HloInstruction *, std::__u::allocator<xla::HloInstruction *>>, std::__u::allocator<std::__u::vector<xla::HloInstruction *, std::__u::allocator<xla::HloInstruction *>>>>::size_type' (aka 'unsigned long') to 'tensorflow::int64' (aka 'long long') in initializer list [-Wc++11-narrowing]
>       {/*x=*/mapping_scheme.GetNumberOfBlocks(), /*y=*/instr_groups.size(),
>                                                        ^~~~~~~~~~~~~~~~~~~
> third_party/tensorflow/compiler/xla/service/gpu/ir_emitter_unnested.cc:3979:56: note: insert an explicit cast to silence this issue
>       {/*x=*/mapping_scheme.GetNumberOfBlocks(), /*y=*/instr_groups.size(),
>                                                        ^~~~~~~~~~~~~~~~~~~
>                                                        static_cast<int64>()
> ```

PR updated. Please help to launch the CI again. Thanks~